@@ -1,90 +1,117 @@
 {
   "name": "jiti",
-  "version": "2.0.0-beta.0",
+  "version": "2.0.0-beta.3",
   "description": "Runtime typescript and ESM support for Node.js",
   "repository": "unjs/jiti",
   "license": "MIT",
+  "type": "module",
   "exports": {
     ".": {
-      "types": "./dist/jiti.d.ts",
-      "default": "./lib/index.js"
-    }
+      "import": {
+        "types": "./lib/jiti.d.mts",
+        "default": "./lib/jiti.mjs"
+      },
+      "require": {
+        "types": "./lib/jiti.d.cts",
+        "default": "./lib/jiti.cjs"
+      }
+    },
+    "./register": {
+      "types": "./lib/jiti-register.d.mts",
+      "import": "./lib/jiti-register.mjs"
+    },
+    "./native": {
+      "bun": "./lib/jiti-native.mjs",
+      "deno": "./lib/jiti-native.mjs",
+      "node": {
+        "import": {
+          "types": "./lib/jiti.d.mts",
+          "default": "./lib/jiti.mjs"
+        },
+        "require": {
+          "types": "./lib/jiti.d.cts",
+          "default": "./lib/jiti.cjs"
+        }
+      },
+      "default": "./lib/jiti-native.mjs"
+    },
+    "./package.json": "./package.json"
   },
-  "main": "./lib/index.js",
-  "types": "./dist/jiti.d.ts",
+  "main": "./lib/jiti.cjs",
+  "module": "./lib/jiti.mjs",
+  "types": "./lib/jiti.d.cts",
   "bin": {
-    "jiti": "bin/jiti.js"
+    "jiti": "./lib/jiti-cli.mjs"
   },
   "files": [
     "lib",
     "dist",
-    "register.js"
+    "register.cjs"
   ],
   "scripts": {
     "build": "pnpm clean && NODE_ENV=production pnpm webpack",
     "clean": "rm -rf dist",
     "dev": "pnpm clean && pnpm webpack --watch",
-    "jiti": "JITI_DEBUG=1 JITI_CACHE=false JITI_REQUIRE_CACHE=false ./bin/jiti.js",
-    "jiti:legacy": "JITI_DEBUG=1 npx node@12 ./bin/jiti.js",
+    "jiti": "JITI_DEBUG=1 ./lib/jiti-cli.mjs",
     "lint": "eslint . && prettier -c src lib test stubs",
     "lint:fix": "eslint --fix . && prettier -w src lib test stubs",
+    "prepack": "pnpm build",
     "release": "pnpm build && pnpm test && changelogen --release --prerelease --push --publish --publishTag 2x",
-    "test": "pnpm lint && vitest run --coverage && pnpm test:bun",
+    "test": "pnpm lint && pnpm test:types && vitest run --coverage && pnpm test:register && pnpm test:bun",
+    "test:types": "tsc --noEmit",
+    "test:register": "node ./test/register-test.mjs",
     "test:bun": "bun --bun test test/bun"
   },
   "devDependencies": {
-    "@babel/core": "^7.24.7",
-    "@babel/helper-module-transforms": "^7.24.7",
-    "@babel/helper-plugin-utils": "^7.24.7",
+    "@babel/core": "^7.25.2",
+    "@babel/helper-module-imports": "^7.24.7",
+    "@babel/helper-module-transforms": "^7.25.2",
+    "@babel/helper-plugin-utils": "^7.24.8",
     "@babel/helper-simple-access": "^7.24.7",
     "@babel/plugin-proposal-decorators": "^7.24.7",
     "@babel/plugin-syntax-class-properties": "^7.12.13",
-    "@babel/plugin-syntax-import-assertions": "^7.24.7",
+    "@babel/plugin-syntax-import-assertions": "^7.25.6",
     "@babel/plugin-transform-export-namespace-from": "^7.24.7",
-    "@babel/plugin-transform-typescript": "^7.24.7",
+    "@babel/plugin-transform-typescript": "^7.25.2",
     "@babel/preset-typescript": "^7.24.7",
-    "@babel/template": "^7.24.7",
-    "@babel/types": "^7.24.7",
+    "@babel/template": "^7.25.0",
+    "@babel/traverse": "^7.25.6",
+    "@babel/types": "^7.25.6",
     "@types/babel__core": "^7.20.5",
+    "@types/babel__helper-module-imports": "^7.18.3",
+    "@types/babel__helper-plugin-utils": "^7.10.3",
     "@types/babel__template": "^7.4.4",
-    "@types/node": "^20.14.9",
-    "@types/object-hash": "^3.0.6",
-    "@types/resolve": "^1.20.6",
-    "@types/semver": "^7.5.8",
-    "@vitest/coverage-v8": "^1.6.0",
-    "acorn": "^8.12.0",
+    "@types/babel__traverse": "^7.20.6",
+    "@types/node": "^22.5.4",
+    "@vitest/coverage-v8": "^2.1.0",
+    "acorn": "^8.12.1",
     "babel-plugin-parameter-decorator": "^1.0.16",
     "babel-plugin-transform-typescript-metadata": "^0.3.2",
     "changelogen": "^0.5.5",
     "config": "^3.3.12",
-    "create-require": "^1.1.1",
     "destr": "^2.0.3",
     "escape-string-regexp": "^5.0.0",
-    "eslint": "^9.5.0",
+    "eslint": "^9.10.0",
     "eslint-config-unjs": "^0.3.2",
-    "esm": "^3.2.25",
     "estree-walker": "^3.0.3",
-    "execa": "^9.3.0",
     "fast-glob": "^3.3.2",
     "mlly": "^1.7.1",
-    "object-hash": "^3.0.0",
     "pathe": "^1.1.2",
-    "pirates": "^4.0.6",
-    "pkg-types": "^1.1.1",
-    "prettier": "^3.3.2",
-    "reflect-metadata": "^0.2.1",
-    "semver": "^7.6.2",
+    "pkg-types": "^1.2.0",
+    "prettier": "^3.3.3",
+    "reflect-metadata": "^0.2.2",
     "std-env": "^3.7.0",
     "terser-webpack-plugin": "^5.3.10",
+    "tinyexec": "^0.3.0",
     "ts-loader": "^9.5.1",
-    "tslib": "^2.6.3",
-    "typescript": "^5.5.2",
-    "unplugin-purge-polyfills": "^0.0.3",
-    "vite": "^5.3.1",
-    "vitest": "^1.6.0",
-    "webpack": "^5.92.1",
+    "typescript": "^5.6.2",
+    "unplugin-purge-polyfills": "^0.0.5",
+    "vitest": "^2.1.0",
+    "webpack": "^5.94.0",
     "webpack-bundle-analyzer": "^4.10.2",
-    "webpack-cli": "^5.1.4"
+    "webpack-cli": "^5.1.4",
+    "webpack-license-plugin": "^4.5.0",
+    "yoctocolors": "^2.1.1"
   },
-  "packageManager": "pnpm@9.4.0"
+  "packageManager": "pnpm@9.10.0"
 }
