@@ -1,6 +1,7 @@
 #!/usr/bin/env node
 
-const { resolve } = require("node:path");
+import { resolve } from "node:path";
+import { createJiti } from "./jiti.cjs";
 
 const script = process.argv.splice(2, 1)[0];
 
@@ -10,8 +11,7 @@ if (!script) {
 }
 
 const pwd = process.cwd();
-const jiti = require("..")(pwd);
+const jiti = createJiti(pwd);
 const resolved = (process.argv[1] = jiti.resolve(resolve(pwd, script)));
 
- 
 jiti.import(resolved).catch(console.error);
