@@ -1,14 +1,21 @@
-import { transformSync } from "@babel/core";
 import type {
   TransformOptions as BabelTransformOptions,
   PluginItem,
 } from "@babel/core";
-import { TransformOptions, TRANSFORM_RESULT } from "./types";
+import { transformSync } from "@babel/core";
+import proposalDecoratorsPlugin from "@babel/plugin-proposal-decorators";
+import syntaxClassPropertiesPlugin from "@babel/plugin-syntax-class-properties";
+import syntaxImportAssertionsPlugin from "@babel/plugin-syntax-import-assertions";
+import transformExportNamespaceFromPlugin from "@babel/plugin-transform-export-namespace-from";
+import transformTypeScriptPlugin from "@babel/plugin-transform-typescript";
+import parameterDecoratorPlugin from "babel-plugin-parameter-decorator";
+import transformTypeScriptMetaPlugin from "babel-plugin-transform-typescript-metadata";
 import { TransformImportMetaPlugin } from "./plugins/babel-plugin-transform-import-meta";
 import { importMetaEnvPlugin } from "./plugins/import-meta-env";
 import transformModulesPlugin from "./plugins/transform-module";
+import type { TransformOptions, TransformResult } from "./types";
 
-export default function transform(opts: TransformOptions): TRANSFORM_RESULT {
+export default function transform(opts: TransformOptions): TransformResult {
   const _opts: BabelTransformOptions & { plugins: PluginItem[] } = {
     babelrc: false,
     configFile: false,
@@ -19,26 +26,33 @@ export default function transform(opts: TransformOptions): TRANSFORM_RESULT {
     cwd: "/",
     ...opts.babel,
     plugins: [
-      [transformModulesPlugin, { allowTopLevelThis: true, async: opts.async }],
+      [
+        transformModulesPlugin,
+        {
+          allowTopLevelThis: true,
+          noInterop: !opts.interopDefault,
+          async: opts.async,
+        },
+      ],
       [TransformImportMetaPlugin, { filename: opts.filename }],
-      [require("@babel/plugin-syntax-class-properties")],
-      [require("@babel/plugin-transform-export-namespace-from")],
+      [syntaxClassPropertiesPlugin],
+      [transformExportNamespaceFromPlugin],
       [importMetaEnvPlugin],
     ],
   };
 
   if (opts.ts) {
     _opts.plugins.push([
-      require("@babel/plugin-transform-typescript"),
+      transformTypeScriptPlugin,
       { allowDeclareFields: true },
     ]);
     // `unshift` because these plugin must come before `@babel/plugin-syntax-class-properties`
     _opts.plugins.unshift(
-      [require("babel-plugin-transform-typescript-metadata")],
-      [require("@babel/plugin-proposal-decorators"), { legacy: true }],
+      [transformTypeScriptMetaPlugin],
+      [proposalDecoratorsPlugin, { legacy: true }],
     );
-    _opts.plugins.push(require("babel-plugin-parameter-decorator"));
-    _opts.plugins.push(require("@babel/plugin-syntax-import-assertions"));
+    _opts.plugins.push(parameterDecoratorPlugin);
+    _opts.plugins.push(syntaxImportAssertionsPlugin);
   }
 
   if (opts.babel && Array.isArray(opts.babel.plugins)) {
