@@ -1,14 +1,18 @@
+import type { Context, JitiResolveOptions } from "./types";
 import { readFileSync } from "node:fs";
 import { builtinModules } from "node:module";
 import { fileURLToPath } from "node:url";
 import { extname } from "pathe";
 import { jitiInteropDefault, normalizeWindowsImportId } from "./utils";
 import { debug } from "./utils";
-import type { Context } from "./types";
 import { jitiResolve } from "./resolve";
 import { evalModule } from "./eval";
 
-export function jitiRequire(ctx: Context, id: string, async: boolean) {
+export function jitiRequire(
+  ctx: Context,
+  id: string,
+  opts: JitiResolveOptions & { async: boolean },
+) {
   const cache = ctx.parentCache || {};
 
   // Check for node: and file: protocol
@@ -20,24 +24,33 @@ export function jitiRequire(ctx: Context, id: string, async: boolean) {
 
   // Check for builtin node module like fs
   if (builtinModules.includes(id) || id === ".pnp.js" /* #24 */) {
-    return ctx.nativeRequire(id);
+    return nativeImportOrRequire(ctx, id, opts.async);
   }
 
   // Experimental Bun support
   if (ctx.opts.experimentalBun && !ctx.opts.transformOptions) {
     try {
-      debug(ctx, `[bun] [native] ${id}`);
-      id = jitiResolve(ctx, id);
-      if (async) {
+      id = jitiResolve(ctx, id, opts);
+      if (!id && opts.try) {
+        return undefined;
+      }
+      debug(
+        ctx,
+        "[bun]",
+        "[native]",
+        opts.async && ctx.nativeImport ? "[import]" : "[require]",
+        id,
+      );
+      if (opts.async && ctx.nativeImport) {
         return ctx.nativeImport(id).then((m: any) => {
-          if (ctx.opts.requireCache === false) {
+          if (ctx.opts.moduleCache === false) {
             delete ctx.nativeRequire.cache[id];
           }
           return jitiInteropDefault(ctx, m);
         });
       } else {
         const _mod = ctx.nativeRequire(id);
-        if (ctx.opts.requireCache === false) {
+        if (ctx.opts.moduleCache === false) {
           delete ctx.nativeRequire.cache[id];
         }
         return jitiInteropDefault(ctx, _mod);
@@ -48,13 +61,16 @@ export function jitiRequire(ctx: Context, id: string, async: boolean) {
   }
 
   // Resolve path
-  const filename = jitiResolve(ctx, id);
+  const filename = jitiResolve(ctx, id, opts);
+  if (!filename && opts.try) {
+    return undefined;
+  }
   const ext = extname(filename);
 
   // Check for .json modules
   if (ext === ".json") {
     debug(ctx, "[json]", filename);
-    const jsonModule = ctx.nativeRequire(id);
+    const jsonModule = ctx.nativeRequire(filename);
     if (jsonModule && !("default" in jsonModule)) {
       Object.defineProperty(jsonModule, "default", {
         value: jsonModule,
@@ -66,21 +82,27 @@ export function jitiRequire(ctx: Context, id: string, async: boolean) {
 
   // Unknown format
   if (ext && !ctx.opts.extensions!.includes(ext)) {
-    debug(ctx, "[unknown]", filename);
-    return nativeImportOrRequire(ctx, id, async);
+    debug(
+      ctx,
+      "[native]",
+      "[unknown]",
+      opts.async ? "[import]" : "[require]",
+      filename,
+    );
+    return nativeImportOrRequire(ctx, filename, opts.async);
   }
 
   // Force native modules
   if (ctx.isNativeRe.test(filename)) {
-    debug(ctx, `[native] ${async ? " [esm]" : ": [cjs]"}`, filename);
-    return nativeImportOrRequire(ctx, id, async);
+    debug(ctx, "[native]", opts.async ? "[import]" : "[require]", filename);
+    return nativeImportOrRequire(ctx, filename, opts.async);
   }
 
-  // Check for CJS cache
+  // Check for runtime cache
   if (cache[filename]) {
     return jitiInteropDefault(ctx, cache[filename]?.exports);
   }
-  if (ctx.opts.requireCache && ctx.nativeRequire.cache[filename]) {
+  if (ctx.opts.moduleCache && ctx.nativeRequire.cache[filename]) {
     return jitiInteropDefault(ctx, ctx.nativeRequire.cache[filename]?.exports);
   }
 
@@ -93,7 +115,7 @@ export function jitiRequire(ctx: Context, id: string, async: boolean) {
     filename,
     ext,
     cache,
-    async,
+    async: opts.async,
   });
 }
 
@@ -102,7 +124,7 @@ export function nativeImportOrRequire(
   id: string,
   async?: boolean,
 ) {
-  return async
+  return async && ctx.nativeImport
     ? ctx
         .nativeImport(normalizeWindowsImportId(id))
         .then((m: any) => jitiInteropDefault(ctx, m))
