@@ -1,32 +1,22 @@
-export type ModuleCache = Record<string, NodeModule>;
-
-export type EvalModuleOptions = Partial<{
-  id: string;
-  filename: string;
-  ext: string;
-  cache: ModuleCache;
-  async?: boolean;
-}>;
-
-export interface JITI extends NodeRequire {
-  transform: (opts: TransformOptions) => string;
-  register: () => () => void;
-  evalModule: (source: string, options?: EvalModuleOptions) => unknown;
-  /** @experimental Behavior of `jiti.import` might change in the future. */
-  import: (id: string) => Promise<unknown>;
-}
-
-type ESMImport = (id: string) => Promise<any>;
+import type { JitiOptions, ModuleCache } from "../lib/types";
+export type {
+  JitiOptions,
+  ModuleCache,
+  EvalModuleOptions,
+  Jiti,
+  TransformOptions,
+  TransformResult,
+  JitiResolveOptions,
+} from "../lib/types";
 
 export interface Context {
   filename: string;
   url: URL;
-  userOptions: JITIOptions;
   parentModule?: NodeModule;
   parentCache?: ModuleCache;
-  nativeImport: ESMImport;
+  nativeImport?: (id: string) => Promise<any>;
   onError?: (error: Error) => void;
-  opts: JITIOptions;
+  opts: JitiOptions;
   nativeModules: string[];
   transformModules: string[];
   isNativeRe: RegExp;
@@ -34,35 +24,5 @@ export interface Context {
   alias?: Record<string, string>;
   additionalExts: string[];
   nativeRequire: NodeRequire;
+  createRequire: (typeof import("node:module"))["createRequire"];
 }
-
-export type TransformOptions = {
-  source: string;
-  filename?: string;
-  ts?: boolean;
-  retainLines?: boolean;
-  async?: boolean;
-  [key: string]: any;
-};
-
-export type TRANSFORM_RESULT = {
-  code: string;
-  error?: any;
-};
-
-export type JITIOptions = {
-  transform?: (opts: TransformOptions) => TRANSFORM_RESULT;
-  debug?: boolean;
-  cache?: boolean | string;
-  sourceMaps?: boolean;
-  requireCache?: boolean;
-  v8cache?: boolean;
-  interopDefault?: boolean;
-  cacheVersion?: string;
-  extensions?: string[];
-  transformOptions?: Omit<TransformOptions, "source">;
-  alias?: Record<string, string>;
-  nativeModules?: string[];
-  transformModules?: string[];
-  experimentalBun?: boolean;
-};
