@@ -4,7 +4,7 @@ import importedWithAssertion from "./file.json" assert { type: "json" };
 
 const required = require("./file.json");
 
-const debug = (label: string, value) =>
+const debug = (label: string, value: any) =>
   console.log(label, ":", value, ".default:", value.default);
 
 debug("Imported", imported);
