@@ -1,3 +1,3 @@
 import fs from "node:fs";
 
-console.log("exists:", fs.existsSync("index.js"));
+console.log("exists:", fs.existsSync(__filename));
